docs: prioritize local storage for theme
I prefer to view doc site in light theme but my OS has dark mode set. Every time I load a page I have to toggle the theme to light because it detects my dark mode preference.

This change makes it so that if a theme preference is already set in localStorage we will use that instead of the OS preference. For folks visiting the first time it will use OS preference. After that it will use whatever they last used.