fix(modal): handle nested modal overflow reset
Resolves https://github.com/uber/baseweb/issues/2435

When closing nested modals simultaneously (e.g. press ESC), the bottom one can unmount before the top one. This causes the top modal to reset the mountNode's overflow to 'hidden', which locks scrolling.
The flow goes like this:
1. Bottom modal mounts, caches last overflow as '', and sets overflow to 'hidden'
2. Top modal mounts, caches last overflow as 'hidden', and sets overflow to 'hidden'
3. Bottom modal unmounts and resets overflow to ''
4. Top modal unmounts and resets overflow to 'hidden'

This change bails out of resetting overflow if it's changed since the modal mounted.
So now the flow would be:
1. Bottom modal mounts, caches last overflow as '', and sets overflow to 'hidden'
2. Top modal mounts, caches last overflow as 'hidden', and sets overflow to 'hidden'
3. Bottom modal unmounts and resets overflow to ''
4. Top modal unmounts and bails out because overflow is no longer
'hidden'