chore(yard): avoid React warning
Avoid React warning:
```
Warning: React does not recognize the `typeDefinition` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `typedefinition` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
```