fix(button): respect focus handler props
Originally I was going to update the Popover "hover" trigger to also
trigger on focus events. Reading through our docs, I noticed that we
already have this feature (under the Accessibility section). Sure
enough, we already pass through the focus events to the anchor (or
wrapping span).

The issue is that Button overwrites any `onFocus` and `onBlur` handlers
passed as top level props. It respects overrides for those props but
technically people can pass `onFocus` and `onBlur` as top level props
so this commit updates the handlers to restore that option.