fix: pass date-io dates into onChange
Letting the user pass in a moment date, and then having a JSDate be
passed to onChange was just a bad idea that was introducing bugs

Returning the generic date-io object meant that any library code that
was assuming the returned object was going to be a JSDate needed
to be made generic as well.

The only instance I could find was these calls in the data-table.
While the ideal solution would be to pass utils into `applyTimeToDate`
, I'm going to hold off on doing that until all of the date functions in
the date picker have been moved over to use utils as well.

Instead I just converted back into a JSDate(which for standard users
won't do anything).