refactor: swap date handler with props
This pulls out of all of the context depdenty date swapping
functionality and uses props instead

This also adds a dedicated DateHelpers class.  When writing
functions that are dependent on the date adapter, it needs
to be passed all over the place, and correctly typed in each
instance.  For Timepicker, this was annoying but not a huge
deal, but for Datepicker, that would be unworkable.
DateHelpers makes the process of writing adapter dependent
functions much easier.