fix: make adapters locale aware
The existing pattern in the date-picker helpers that is being recreated is based
on passing locale as an argument into  functions and having that update
the locale, but for date-io, their paradigm is to pass the locale in the adapter
when it's created.  This function makes these two strategies play nice
together.

When locale is needed in a function, getAdapterWithNewLocale recreates the
adapter for the new locale, and everything works.