chore: confirm index and date-helpers match
This adds a new wrapper around helpers in the tests that
for every tests that runs, runs the function in both date-helpers and
utils/index and confirms that they return the same value

This will helps ensure that there's parity between the two
interfaces.

Down the road, this could also be adapted to test the other utils
adapters.