fix: remove differenceInCalendarDays
When I migrated differenceInCalendarDays over from date-fns
I didn't fully capturing it's behavior.  The date-fns implementation
was doing stuff with Date.getTimezoneOffset which would have
been way to complicated in the `date-io` api.  So, to try to simplify
it, I just got the difference in milliseconds and divided it by the
number of milliseconds in a day.

This was just a bad idea all around.  The number of ms in a day
can change, and it wasn't even the exact behavior we wanted.
If we needed to  recreate the differenceInCalendarDays behavior
exactly, we'd actually be kind of in trouble, luckily it was only being
used to figure out if a date was on or after/before a given
date(mostly for dealing with minDate and maxDate).

So, instead of trying to re-implement, i just took it out all-together
and replaced it with isOnOrAfterDay and isOnOrBeforeDay.

I also added some test cases that catch the problem cases that
slipped through before.