Fix broken Card typescript types
The index.d.ts for the Card component was importing types from a flow types file causing a typescript error when importing from `baseui/card`:

```
error TS7016: Could not find a declaration file for module './types.js'. '/Users/airhorns/Code/gadget/node_modules/baseui/card/types.js' implicitly has an 'any' type.

3 import {ImagePropsT} from './types.js';
```

The type for the `headerImage` prop is a string or all the props that an `<img>` can take, so this updates the TypeScript type to just use React's built in props list for images.