feat(select): add a controlRef prop to <Select/>
Similar to <Input/>, it's useful to be able to get a ref to the inner element so developers can imperatively do stuff to it, like focus it. This adds a `controlRef` prop to <Select/> so you can do just that!

It's called `controlRef` and not `inputRef`, which is different than <Input/>, but on purpose, because if the Select isn't searchable, then there isn't actually an `<input>` created, just a `<div>` as a container. The Flow and TypeScript types reflect this, but I thought it might be good to make the name and docs clear that it won't always be an input.

Added a test and docs.

Fixes #3398